% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanfield_functions.R
\name{MFA_sim}
\alias{MFA_sim}
\title{Computes mean-field solutions}
\usage{
MFA_sim(
  alpha_sim = NA,
  beta_sim = NA,
  average_density = 2,
  MFA_to_solve = c("Ising", "Percolation", "han", "zero_corrected", "independent_a",
    "Blume_capel", "Potts", "Continuous_01")
)
}
\arguments{
\item{alpha_sim}{Numeric vector, external field parameter space.}

\item{beta_sim}{Numeric vector, beta parameter space.}

\item{average_density}{Numeric, average density for the model.}

\item{MFA_to_solve}{Character vector, specifies the mean-field model to solve.
Options: "Ising", "Percolation", "han", "zero_corrected", "independent_a", "Blume_capel", "Potts", "Continuous_01".}
}
\value{
A list or data frame with the simulation results.
}
\description{
This function computes the mean-field solutions for various models as specified. It allows for the customization
of the beta parameter space, external field parameter space, and the selection of the mean-field model to solve.
Valid model options include "Ising", "Percolation", "han", "zero_corrected", "independent_a", "Blume_capel",
"Potts", and "Continuous_01".
}
\examples{
MFA_sim(alpha_sim = seq(-3,3, length.out = 100), beta_sim = seq(0,3, length.out = 100), average_density = 2, MFA_to_solve = "Ising")
}
