% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanfield_functions.R
\name{MFA_solver}
\alias{MFA_solver}
\title{Internal function that helps solving the mean field}
\usage{
MFA_solver(
  root_MFA,
  alpha_sim = 0,
  beta_sim = 1,
  average_density = 4,
  MFA_to_solve = MFA_to_solve
)
}
\arguments{
\item{root_MFA}{Function, the root mean-field function to be solved.}

\item{alpha_sim}{Numeric, external field parameter space.}

\item{beta_sim}{Numeric, beta parameter space.}

\item{average_density}{Numeric, average density for the model.}

\item{MFA_to_solve}{Character, specifies the mean-field model being solved.}
}
\value{
A data frame with the simulation results, including stability and value of each root found.
}
\description{
This internal function assists in solving the mean field by employing simulations based on specified parameters.
It is primarily used within the MFA_sim function.
}
\examples{
# Example usage not provided as this function is intended for internal use by MFA_sim()
}
