% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanfield_functions.R
\name{plot_MFA}
\alias{plot_MFA}
\title{Plots a MFA_sim() object}
\usage{
plot_MFA(
  result,
  camera_x = 2,
  camera_y = 0,
  camera_z = 0,
  y_label = "Alpha",
  x_label = "Beta",
  z_label = "Mean field"
)
}
\arguments{
\item{result}{The output of the MFA_sim() function, expected to be a data frame.}

\item{camera_x}{Numeric, controls the camera angle of the x-axis.}

\item{camera_y}{Numeric, controls the camera angle of the y-axis.}

\item{camera_z}{Numeric, controls the camera angle of the z-axis.}

\item{y_label}{Character, label of the y-axis.}

\item{x_label}{Character, label of the x-axis.}

\item{z_label}{Character, label of the z-axis.}
}
\value{
A 3D plotly object.
}
\description{
This function creates a 3D plot of the results from the MFA_sim() function. It allows customization of camera angles
and axis labels for detailed visualization.
}
\examples{
result <- MFA_sim(alpha_sim = seq(-3,3, length.out = 100), beta_sim = seq(0,3, length.out = 100), average_density = 2, MFA_to_solve = "Ising")
plot_MFA(result)

}
