% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanfield_functions.R
\name{save_plot_MFA}
\alias{save_plot_MFA}
\title{Saves the output of plot_MFA() as a PDF file}
\usage{
save_plot_MFA(MFA_plot, saved_image_name, width = 600, height = 600)
}
\arguments{
\item{MFA_plot}{The plotly object output from \code{plot_MFA()} to be saved.}

\item{saved_image_name}{Name of the saved file ending with ".pdf".}

\item{width}{Numeric, the width of the saved plot in pixels. Defaults to 600.}

\item{height}{Numeric, the height of the saved plot in pixels. Defaults to 600.}
}
\description{
This function saves the 3D plot generated by \code{plot_MFA()} as a PDF file. It requires the setup of a Python environment
using \code{reticulate}. Before using this function, ensure Python is configured by running \code{reticulate::use_miniconda('r-reticulate')}
and \code{reticulate::py_run_string("import sys")}. These steps are necessary to enable R to interface with Python for saving
the plot because the plot saving functionality relies on Python libraries.
}
\examples{
# First, generate a plot with plot_MFA() function.
result <- MFA_sim(alpha_sim = seq(-3,3, length.out = 100), beta_sim = seq(0,3, length.out = 100), average_density = 2, MFA_to_solve = "Ising")
MFA_plot <- plot_MFA(result)
# Then, save the generated plot as a PDF file.
reticulate::use_miniconda('r-reticulate')
reticulate::py_run_string("import sys")
save_plot_MFA(MFA_plot)
}
